journal seperated in modules
