+++ journal.html
@@ -48,7 +48,7 @@
 
         <div id="comment-overlay" class="fixed z-10 top-0 bottom-0 w-dvw min-h-dvh bg-black/50 hidden"></div>
 
-    <script src="journal.js"></script>
+    <script type="module" src="journal.js"></script>
 </body>
 
 </html>
